requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
questionary>=2.0.0